refactor(cli): decouple Ethereum-specific logic from parser_cli
Move Ethereum-specific code (ABI registry, network metadata) into
ethereum.rs and Solana-specific code (IDL mappings, Solana metadata)
into solana.rs. cli.rs now dispatches to chain modules with no direct
visualsign_ethereum imports.

Closes #186. Prerequisite for #187.
